requirements.txt
absl-py==2.2.2
actionlib==1.14.0
addict==2.4.0
aenum==3.1.16
aiobotocore==2.12.1
aiodns==3.1.1
aiofiles==23.2.1
aiohttp==3.9.4
aioitertools==0.11.0
aiosignal==1.3.1
ale-py==0.11.0
angles==1.9.13
annotated-types==0.7.0
anyio==4.4.0
asteval==0.9.32
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.28.0
azure-identity==1.13.0
azure-storage-blob==12.17.0
bddl==3.5.0
boto3==1.26.63
botocore==1.34.68
Brotli @ file:///croot/brotli-split_1736182456865/work
camera-calibration-parsers==1.12.0
catkin==0.8.10
cchardet==2.1.7
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
construct==2.10.68
contourpy==1.3.2
coverage==7.4.4
cryptography==41.0.7
cv-bridge==1.16.2
cycler==0.11.0
decorator==5.2.1
diagnostic-updater==1.11.0
dm-tree==0.1.9
docker-pycreds==0.4.0
dynamic-reconfigure==1.7.3
exceptiongroup==1.2.1
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.110.0
fastjsonschema==2.21.1
filelock @ file:///croot/filelock_1744281381737/work
fonttools==4.57.0
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.3
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///croot/gmpy2_1738085463648/work
graphviz==0.20.3
grpcio==1.71.0
gunicorn==22.0.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
idna==3.7
idna-ssl==1.1.0
ImageHash==4.3.2
imageio==2.33.1
imageio-ffmpeg==0.4.9
iniconfig==2.1.0
interactive-markers==1.12.0
ipython==8.20.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
jupytext==1.17.1
kiwisolver==1.4.4
laser_geometry==1.6.7
llvmlite==0.42.0
loguru==0.7.3
lxml==5.2.2
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
message-filters==1.17.0
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath==1.3.0
msal==1.23.0
msal-extensions==1.0.0
msgpack==1.1.0
multidict==6.0.5
nbformat==5.10.4
nest-asyncio==1.5.9
networkx==3.2.1
nltk==3.9.1
numba==0.59.1
numpy==1.26.0
numpy-quaternion==2023.0.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-lula-no-cuda==0.10.1
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
nvidia_srl_base==0.10.0
nvidia_srl_math==0.9.0
nvidia_srl_usd==0.14.0
nvidia_srl_usd_to_urdf==0.6.0
nvsmi==0.4.2
oauthlib==3.2.2
omnigibson==1.1.1
opencv-python==4.9.0.80
ormsgpack==1.7.0
osqp==0.6.5
packaging==23.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
Pint==0.20.1
platformdirs==4.3.7
plotly==5.3.1
pluggy==1.5.0
portalocker==2.7.0
progressbar==2.5
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycares==4.3.0
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
pygame==2.6.1
Pygments==2.19.1
pymeshlab==2022.2.post4
pyparsing==3.0.9
pyperclip==1.8.0
pypng==0.20220715.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==8.3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-qt-binding==0.4.4
pytz==2024.1
PyWavelets==1.8.0
PyYAML==6.0.1
qdldl==0.1.7.post1
qrcode==7.4.2
ray==2.44.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rich==14.0.0
rosbag==1.17.0
rosclean==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslz4==1.17.0
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rpds-py==0.24.0
Rtree==1.2.0
rviz==1.14.25
s3transfer==0.6.1
scipy==1.10.1
selenium==4.14.0
sensor-msgs==1.13.1
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools-scm==8.0.4
shapely==2.1.0
Shimmy==2.0.0
six==1.16.0
smclib==1.8.6
smmap==5.0.2
sniffio==1.3.1
stable_baselines3==2.6.0
stack-data==0.6.3
starlette==0.36.3
svgwrite==1.4.3
sympy==1.12
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
tf==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
toml==0.10.2
tomli==2.0.1
topic-tools==1.17.0
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tornado==6.2
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
Tree==0.2.4
trimesh==4.0.10
triton==3.1.0
typing_extensions==4.10.0
tzdata==2025.2
urllib3==2.2.2
uvicorn==0.29.0
wandb==0.19.10
watchdog==4.0.0
wcwidth==0.2.13
webbot==0.34
websockets==12.0
Werkzeug==3.1.3
wrapt==1.16.0
xacro==1.14.18
yarl==1.9.4
